+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateFileOperateService。
   - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateDataInfoOperateService。
   - com.dwarfeng.familyhelper.life.stack.service.ActivityOperateService。
+  - com.dwarfeng.familyhelper.life.stack.service.ActivityCoverOperateService。
 
 - 新增实体及其维护服务，单元测试通过。
   - com.dwarfeng.familyhelper.life.stack.bean.entity.Activity。